📦 refactor: rename status actions.ts to data.ts
- getStatus is called directly during Server Component rendering (info.tsx), not invoked from client-side code as a mutation, so it isn't a real Next.js Server Action despite the 'actions.ts' name
- data.ts follows the convention of reserving 'actions.ts'/'use server' for mutations, keeping server-side reads separate
